unread count

+++ Spika/UIFlow/MainFlow/Chats/AllChats/View/AllChatsTableViewCell.swift
@@ -75,7 +75,7 @@ class AllChatsTableViewCell: UITableViewCell, BaseView {
 //        mutedIcon.constrainWidth(20)
     }
     
-    func configureCell(avatarUrl: URL?, name: String, description: String, time: String, badgeNumber: Int, muted: Bool, pinned: Bool) {
+    func configureCell(avatarUrl: URL?, name: String, description: String, time: String, badgeNumber: Int64, muted: Bool, pinned: Bool) {
         leftImageView.kf.setImage(with: avatarUrl, placeholder: UIImage(safeImage: .userImage))
 
         nameLabel.text = name
